chore: update readme and add script to create delegation
